requirements.txt
psycopg2-binary==2.9.11
pymongo==4.15.5
neo4j==6.0.3
redis==7.1.0
python-dotenv==1.0.0

